change path (renaming)  from tajo to tajo.jdbc package
