+++ tajo-jdbc/pom.xml
@@ -25,10 +25,10 @@
   </parent>
 
   <groupId>org.apache.zeppelin</groupId>
-  <artifactId>zeppelin-tajo</artifactId>
+  <artifactId>zeppelin-tajo-jdbc</artifactId>
   <packaging>jar</packaging>
   <version>0.5.0-incubating-SNAPSHOT</version>
-  <name>Zeppelin: Tajo interpreter</name>
+  <name>Zeppelin: Tajo JDBC interpreter</name>
   <url>http://www.apache.org</url>
 
   <properties>
@@ -100,7 +100,7 @@
               <goal>copy-dependencies</goal>
             </goals>
             <configuration>
-              <outputDirectory>${project.build.directory}/../../interpreter/tajo</outputDirectory>
+              <outputDirectory>${project.build.directory}/../../interpreter/tajo-jdbc</outputDirectory>
               <overWriteReleases>false</overWriteReleases>
               <overWriteSnapshots>false</overWriteSnapshots>
               <overWriteIfNewer>true</overWriteIfNewer>
@@ -114,7 +114,7 @@
               <goal>copy</goal>
             </goals>
             <configuration>
-              <outputDirectory>${project.build.directory}/../../interpreter/tajo</outputDirectory>
+              <outputDirectory>${project.build.directory}/../../interpreter/tajo-jdbc</outputDirectory>
               <overWriteReleases>false</overWriteReleases>
               <overWriteSnapshots>false</overWriteSnapshots>
               <overWriteIfNewer>true</overWriteIfNewer>
