+++ tajo-jdbc/src/test/java/org/apache/zeppelin/tajo/jdbc/TesterResultSet.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.zeppelin.tajo;
+package org.apache.zeppelin.tajo.jdbc;
 
 import java.math.BigDecimal;
 import java.sql.*;
